Fix setup.py so that install via pip works
Create a src/github symbolic link that points to the src/ directory so
that distutils can find the files.

http://docs.python.org/distutils/setupscript.html

"""
If you use a different convention to lay out your source directory,
that's no problem: you just have to supply the package_dir option to
tell the Distutils about your convention. For example, say you keep all
Python source under lib, so that modules in the "root package"
(i.e., not in any package at all) are in lib, modules in the foo package
are in lib/foo, and so forth. Then you would put

  package_dir = {'': 'lib'}

in your setup script.  The keys to this dictionary are package names,
and an empty package name stands for the root package. The values are
directory names relative to your distribution root. In this case, when
you say packages = ['foo'], you are promising that the file
lib/foo/__init__.py exists.
"""